backend/requirements.txt
google-genai
faiss-cpu
langchain
langchain-community
streamlit
python-dotenv
fastapi
uvicorn[standard]